refactor: Cambiato namespace lib/freeRTOS a `freertos`
