+++ app/lib/freeRTOS/Mutex.cpp
@@ -1,26 +1,26 @@
 #include "Mutex.h"
 #include <cassert>
 
-mcu::Mutex::Mutex() : 
+freertos::Mutex::Mutex() : 
     m_handle(xSemaphoreCreateMutexStatic(&m_buffer))
 {
     assert(m_handle != NULL);
 }
 
-void mcu::Mutex::lock() {
+void freertos::Mutex::lock() {
     bool result = xSemaphoreTake(m_handle, portMAX_DELAY) == pdTRUE;
     assert(result);
 }
 
-bool mcu::Mutex::try_lock() {
+bool freertos::Mutex::try_lock() {
     return xSemaphoreTake(m_handle, 0) == pdTRUE;
 }
 
-void mcu::Mutex::unlock() {
+void freertos::Mutex::unlock() {
     bool result = xSemaphoreGive(m_handle) == pdTRUE;
     assert(result);
 }
 
-mcu::Mutex::~Mutex() {
+freertos::Mutex::~Mutex() {
     vSemaphoreDelete(m_handle);
 }
